+++ app/src/main/java/com/dropdrage/simpleweather/domain/location/Location.kt
@@ -1,3 +1,3 @@
 package com.dropdrage.simpleweather.domain.location
 
-data class Location(val latitude: Double, val longitude: Double)
+data class Location(val latitude: Float, val longitude: Float)
